requirements.txt
astroid==2.0.4
autopep8==1.4
backcall==0.1.0
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
decorator==4.3.2
Django==2.2.4
django-debug-toolbar==1.11
django-extensions==2.1.7
django-jalali==3.1.0
django-taggit==1.1.0
django-webpack-loader==0.6.0
djangorestframework==3.10.3
idna==2.8
ipython==7.2.0
ipython-genutils==0.2.0
isort==4.3.4
jdatetime==3.0.2
jedi==0.13.2
lazy-object-proxy==1.3.1
Markdown==3.1
mccabe==0.6.1
parso==0.3.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.7
psycopg2==2.8.2
psycopg2-binary==2.7.5
ptyprocess==0.6.0
pycodestyle==2.4.0
Pygments==2.3.1
PyJWT==1.7.1
pylint==2.1.1
pytz==2018.5
requests==2.22.0
rope==0.11.0
six==1.11.0
sorl-thumbnail==12.5.0
soupsieve==1.7.3
sqlparse==0.2.4
traitlets==4.3.2
typed-ast==1.1.0
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.10.11
